Add clearer messaging for under-indented lines (#8)
Add a clearer message than “deficient indentation” when a line is under-indented, including how much the line *should* be indented. The indentation warning now looks like:

    example.yaml
      5:2        warning  line is under-indented (it should be indented at least 5)

It doesn’t show whether spaces or tabs should be used since we don’t actually know what kind of indentation the file uses. This is basically the same as where we assume spaces when fixing the problem. Both should be fixed together as a separate issue.

Fixes #1.