Prototype support for file globs
This is a bit ugly and needs to be refactored, but it's a basic implementation of support for running yaml-doctor agains multiple paths, directories, and globs. Any positional arguments are treated as glob paths. If a path is actually a directory, it will be treated like `path/**/*.{yaml,yml,md}`.

Fixes #2.