November 2022 Update & Cleanup (#24)
This is a basic cleanup that updates dependencies (to fix some security issues) and adds current versions of Node.js to the test matrix. It’s a breaking change, since it moves our minimum required Node.js version from v8 to v12 (which is also now specified in package.json).

Supersedes #17.