Upgrade glob to v9
This upgrades `glob` to the latest of the v9 line and updates our minimum Node.js version to 16 (from 12). It's a little complicated, since glob and some of its dependencies have updated their Node.js requirements in patch releases, which is why there's the added condition on lru-cache (not a new dependency, just a transitive one that needs more narrow control).

We could also update to Node.js v18 and not worry about all this, but I want to do the minimum bump to clean up scary warnings and issues when you install right now, and we only need Node.js 16 for that.