Handle globs that match a directory
When you ask YAML-Doctor to check a directory (or directories), it looks inside the directory for YAML files:

```
yaml-doctor path/to/directory
yaml-doctor path/to/directory-pattern*
```

But if you pass a directory-matching pattern directly (instead of having your shell interpret it), we fail:

```
yaml-doctor 'path/to/directory-pattern*'
```

This solves the issue by handling directories *while* globbing instead of before globbing.

Fixes #30.